--- server/module-domain/src/main/java/com/mile/comment/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.mile.comment.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
