--- server/module-domain/src/main/java/com/mile/comment/domain/BaseTimeEntity.java
@@ -1,14 +1,0 @@
-package com.mile.comment.domain;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeEntity {
-    @CreatedDate
-    private LocalDateTime createdAt;
-}
