+++ server/module-domain/src/main/java/com/mile/comment/domain/Comment.java
@@ -1,5 +1,6 @@
 package com.mile.comment.domain;
 
+import com.mile.config.BaseTimeEntity;
 import com.mile.post.domain.Post;
 import com.mile.user.domain.User;
 import jakarta.persistence.Entity;
